Rename main_app and mini to satellite and minisat
