+++ satellite/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class MainAppConfig(AppConfig):
-    name = 'main_app'
+    name = 'satellite'
