+++ satellite/modules/dashboard_details.py
