+++ satellite/modules/docker_manage.py
