+++ satellite/modules/kickstart.py
