+++ satellite/modules/vm_manage.py
