+++ satellite/static/myscript.js
