+++ satellite/static/vmlist.js
