+++ satellite/templates/VM_info.html
