+++ satellite/templates/containers/Dockerfile
