+++ satellite/templates/host/create_host.html
