+++ satellite/templates/host/operating_system.html
