+++ satellite/templates/infrastructure/compute_resource.html
