+++ satellite/templatetags/extra_tags.py
